requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bitarray==3.0.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cytoolz==1.0.0
discord==2.3.2
discord.py==2.4.0
eth-abi==4.2.1
eth-account==0.10.0
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==1.0.1
eth-typing==5.0.1
eth-utils==2.3.2
Flask==3.0.3
frozenlist==1.5.0
hexbytes==0.3.1
hyperliquid-python-sdk==0.9.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
parsimonious==0.9.0
propcache==0.2.0
pycryptodome==3.21.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
toolz==1.0.0
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
